removed unnecessary classes for better test coverage
