+++ src/main/java/com/websummarizer/Web/Summarizer/controller/ShortLinkGenerator.java
@@ -29,25 +29,4 @@ public class ShortLinkGenerator {
         }
         return out.toString();
     }
-
-    //unfinished checkpassword function
-    public boolean checkPassword(String password){
-        String pattern = "^(?=.*[A-Z])(?=.*[a-z])(?=.*\\d)(?=.*[@$!%*?&])[A-Za-z\\d@$!%*?&]{8,}$";
-
-        // Compile the pattern
-        Pattern regex = Pattern.compile(pattern);
-
-        // Create a Matcher object
-        Matcher matcher = regex.matcher(password);
-        return matcher.matches();
-    }
-
-//    //check password
-//        if(!shortlink.checkPassword(password)) {
-//        model.addAttribute("isValid", false);
-//        model.addAttribute("html", "<span class=\"bi bi-exclamation-triangle-fill\"></span>");
-//        model.addAttribute("message", "Password must contain at least 8 characters, 1 uppercase letter, 1 lowercase letter, 1 number, and 1 special character");
-//
-//        return "user/register";
-//    }
 }
