--- src/main/java/com/websummarizer/Web/Summarizer/model/BitlyRequest.java
@@ -1,9 +1,0 @@
-package com.websummarizer.Web.Summarizer.model;
-
-import lombok.Data;
-
-@Data
-public class BitlyRequest {
-
-    private String longURL;
-}
