+++ src/main/resources/env.properties
@@ -1,27 +1,27 @@
 #API CREDENTIALS:
 #BART
-API_AUTH_TOKEN = ##
+API_AUTH_TOKEN = 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
 #OPEN AI
-API_AUTH_TOKEN_OPEN_AI = ##
+API_AUTH_TOKEN_OPEN_AI = sk-proj-xWHIhK0X1AJAy8GdAqgYT3BlbkFJU2wVvtbh5zMLUAQJb5V0
 
 #DB CREDENTIALS
-DB_USERNAME_UAT = ##
-DB_PASSWORD_UAT = ##
+DB_USERNAME_UAT = postgres
+DB_PASSWORD_UAT = WebSummarizer786@
 
 #GITHUB CREDENTIALS
-GITHUB_CLIENT_ID = ##
-GITHUB_CLIENT_SECRET = ##
+GITHUB_CLIENT_ID = 327c918f7f06045a5816
+GITHUB_CLIENT_SECRET = 750cda4d9c6739fc0c1a2b4c0e11334116bc24d9
 
 #GOOGLE CREDENTIALS
-GOOGLE_CLIENT_ID = ##
-GOOGLE_CLIENT_SECRET = ##
+GOOGLE_CLIENT_ID = 51331619550-vo0ogu79jqjhusc6pc7892q3lakp2rmm.apps.googleusercontent.com
+GOOGLE_CLIENT_SECRET = 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
 
 #FACEBOOK CREDENTIAL
-FACEBOOK_CLIENT_ID = ##
-FACEBOOK_CLIENT_SECRET = ##
+FACEBOOK_CLIENT_ID = FACEBOOK_CLIENT_ID
+FACEBOOK_CLIENT_SECRET = FACEBOOK_CLIENT_SECRET
 
 #SMTP CREDENTIALS GMAIL
-SMTP_USERNAME = ##
-SMTP_PASSWORD =  ##
+SMTP_USERNAME = wsummarizer@gmail.com
+SMTP_PASSWORD =  WebSummarizer786@
 
 BITLY_AUTH_TOKEN = ##
